requirements.txt
anyio==3.6.2
argcomplete==3.0.8
attrs==23.1.0
bcrypt==4.0.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
datamodel-code-generator==0.18.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
fastapi-restful==0.4.3
genson==1.2.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflect==5.6.2
intasend-python==1.0.7
isort==5.11.5
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==23.1
passlib==1.7.4
pathable==0.4.3
pathspec==0.11.1
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
prance==0.22.11.4.0
psutil==5.9.5
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.3
pyrsistent==0.19.3
PySnooper==1.1.1
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.30.0
rsa==4.9
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
semver==2.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
zipp==3.15.0
